backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.3.1
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
playwright>=1.48.0
groq>=0.11.0
faker==30.0.0
python-multipart==0.0.12
aiofiles==24.1.0
Pillow==10.4.0
websockets==13.1
httpx>=0.28.1
tenacity==8.5.0
jinja2==3.1.4
markdown==3.7
